Project and JSON restructure
