+++ data/osa_slips.json
@@ -1,0 +1,44 @@
+[
+    {
+        "name": "Benny",
+        "email": "2233116@slu.edu.ph",
+        "absences": [
+            {
+                "date": "2024-11-28",
+                "reason": "Sickeness",
+                "course": "9377",
+                "status": "Pending"
+            },
+            {
+                "date": "",
+                "reason": "test",
+                "course": "test",
+                "status": "Pending"
+            },
+            {
+                "date": "",
+                "reason": "test",
+                "course": "test",
+                "status": "Pending"
+            },
+            {
+                "date": "",
+                "reason": "s",
+                "course": "s",
+                "status": "Pending"
+            },
+            {
+                "date": "",
+                "reason": "s",
+                "course": "s",
+                "status": "Pending"
+            },
+            {
+                "date": "",
+                "reason": "",
+                "course": "",
+                "status": "pending"
+            }
+        ]
+    }
+]
