+++ osa/routes.py
@@ -1,0 +1,76 @@
+from flask import Flask, request, jsonify, render_template, redirect, url_for, session
+from osa.services import osa_system
+
+app = Flask(__name__)
+app.secret_key = 'secret_key'
+
+@app.route('/')
+def index():
+    return render_template('index.html')
+
+@app.route('/login', methods=['POST'])
+def login():
+    data = request.json
+    email = data.get('email')
+    password = data.get('password')
+
+    user = osa_system.validate_user(email, password)
+    if user:
+        session['email'] = user.email
+        session['usertype'] = user.usertype
+        if user.usertype == 'admin':
+            return jsonify({'redirect': '/admin'})
+        elif user.usertype == 'student':
+            return jsonify({'redirect': '/student'})
+    return jsonify({'error': 'Invalid email or password'}), 401
+
+@app.route('/admin')
+def admin():
+    if 'usertype' in session and session['usertype'] == 'admin':
+        return render_template('admin.html')
+    return redirect(url_for('index'))
+
+@app.route('/student')
+def student():
+    if 'usertype' in session and session['usertype'] == 'student':
+        return render_template('student.html')
+    return redirect(url_for('index'))
+
+@app.route('/add_student', methods=['POST'])
+def add_student():
+    if 'usertype' in session and session['usertype'] == 'student':
+        data = request.json
+        required_fields = ['date_absent', 'reason', 'course']
+        if not all(field in data for field in required_fields):
+            return jsonify({"error": "Missing required fields"}), 400
+
+        email = session['email']
+        student = next((s for s in osa_system.students if s.email == email), None)
+        if student:
+            name = student.name
+            date_absent = data['date_absent']
+            reason = data['reason']
+            course = data['course']
+            osa_system.add_student(name, email, date_absent, reason, course)
+            results = osa_system.process_student(student)
+            return jsonify(results)
+    return jsonify({"error": "Unauthorized"}), 401
+
+@app.route('/osaform')
+def osaform():
+    return render_template('osaform.html')
+
+@app.route('/history')
+def history():
+    if 'usertype' in session and session['usertype'] == 'student':
+        return render_template('history.html')
+    return redirect(url_for('index'))
+
+@app.route('/api/absences')
+def api_absences():
+    if 'usertype' in session and session['usertype'] == 'student':
+        email = session['email']
+        student = next((s for s in osa_system.students if s.email == email), None)
+        if student:
+            return jsonify(student.absences)
+    return jsonify([]), 404
