+++ osa/templates/history.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>OSA Slip History</title>
+    <link rel="stylesheet" href="static/css/history.css">
+    <script src="static/js/history.js"></script>
+</head>
+<body>
+    <div class="title-container">
+        <h1>OSA Slip History</h1>
+        <button onclick="location.href='/student'">Back</button>
+    </div>
+    <div id="history-container"></div>
+</body>
+</html>
