+++ osa/templates/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>OSA Slip System - Login</title>
+    <link rel="stylesheet" href="static/css/index.css">
+    <script src="static/js/index.js"></script>
+</head>
+<body>
+<h1>Login to OSA Slip System</h1>
+<form id="loginForm">
+    <label for="email">Email:</label>
+    <input type="email" id="email" name="email" required><br>
+    <label for="password">Password:</label>
+    <input type="password" id="password" name="password" required><br>
+    <button type="button" onclick="submitLogin()">Login</button>
+</form>
+<div id="results"></div>
+</body>
+</html>
