+++ osa/templates/osaform.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>OSA Slip Distribution</title>
+    <link rel="stylesheet" href="static/css/osaform.css">
+    <script src="static/js/osaform.js"></script>
+</head>
+<body>
+    <h1>Automated OSA Slip Distribution</h1>
+    <form id="osaForm">
+        <label for="date_absent">Date of Absence (YYYY-MM-DD):</label>
+        <input type="date" id="date_absent" name="date_absent" required><br>
+        <label for="reason">Reason for Absence:</label>
+        <input type="text" id="reason" name="reason" required><br>
+        <label for="course">Course Code or Name:</label>
+        <input type="text" id="course" name="course" required><br>
+        <div class="form-buttons">
+            <button type="button" onclick="submitForm()">Submit</button>
+            <button type="button" onclick="goBack()">Go Back</button>
+        </div>
+    </form>
+    <div id="results"></div>
+</body>
+</html>
