+++ osa/templates/student.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>OSA Slip System</title>
 
-    <link rel="stylesheet" href="static/student.css">
+    <link rel="stylesheet" href="static/css/student.css">
 </head>
 <body>
     <h1>Hello <span id="name">name</span></h1>
