--- templates/history.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>OSA Slip History</title>
-    <link rel="stylesheet" href="static/history.css">
-</head>
-<body>
-    <div class="title-container">
-        <h1>OSA Slip History</h1>
-        <button onclick="location.href='/student'">Back</button>
-    </div>
-    <div id="history-container"></div>
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            fetch('/api/absences')
-                .then(response => response.json())
-                .then(data => {
-                    const container = document.getElementById('history-container');
-                    data.forEach(absence => {
-                        const card = document.createElement('div');
-                        card.className = 'card';
-                        card.innerHTML = `
-                            <h2>${absence.date}</h2>
-                            <p><strong>Reason:</strong> ${absence.reason}</p>
-                            <p><strong>Course:</strong> ${absence.course}</p>
-                        `;
-                        container.appendChild(card);
-                    });
-                });
-        });
-    </script>
-</body>
-</html>
