--- templates/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>OSA Slip System - Login</title>
-    <link rel="stylesheet" href="static/index.css">
-</head>
-<body>
-    <h1>Login to OSA Slip System</h1>
-    <form id="loginForm">
-        <label for="email">Email:</label>
-        <input type="email" id="email" name="email" required><br>
-        <label for="password">Password:</label>
-        <input type="password" id="password" name="password" required><br>
-        <button type="button" onclick="submitLogin()">Login</button>
-    </form>
-    <div id="results"></div>
-    <script>
-        function submitLogin() {
-            const formData = {
-                email: document.getElementById('email').value,
-                password: document.getElementById('password').value
-            };
-            fetch('/login', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify(formData)
-            })
-            .then(response => response.json())
-            .then(data => {
-                const resultsDiv = document.getElementById('results');
-                resultsDiv.innerHTML = '';
-                if (data.error) {
-                    const p = document.createElement('p');
-                    p.textContent = data.error;
-                    resultsDiv.appendChild(p);
-                } else if (data.redirect) {
-                    window.location.href = data.redirect;
-                }
-            });
-        }
-    </script>
-</body>
-</html>
