--- templates/osaform.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>OSA Slip Distribution</title>
-    <link rel="stylesheet" href="static/osaform.css">
-</head>
-<body>
-    <h1>Automated OSA Slip Distribution</h1>
-    <form id="osaForm">
-        <label for="date_absent">Date of Absence (YYYY-MM-DD):</label>
-        <input type="date" id="date_absent" name="date_absent" required><br>
-        <label for="reason">Reason for Absence:</label>
-        <input type="text" id="reason" name="reason" required><br>
-        <label for="course">Course Code or Name:</label>
-        <input type="text" id="course" name="course" required><br>
-        <div class="form-buttons">
-            <button type="button" onclick="submitForm()">Submit</button>
-            <button type="button" onclick="goBack()">Go Back</button>
-        </div>
-    </form>
-    <div id="results"></div>
-    <script>
-        function submitForm() {
-            const formData = {
-                date_absent: document.getElementById('date_absent').value,
-                reason: document.getElementById('reason').value,
-                course: document.getElementById('course').value
-            };
-            fetch('/add_student', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify(formData)
-            })
-            .then(response => response.json())
-            .then(data => {
-                const resultsDiv = document.getElementById('results');
-                resultsDiv.innerHTML = '';
-                if (data.error) {
-                    const p = document.createElement('p');
-                    p.textContent = data.error;
-                    resultsDiv.appendChild(p);
-                } else {
-                    data.forEach(result => {
-                        const p = document.createElement('p');
-                        p.textContent = result;
-                        resultsDiv.appendChild(p);
-                    });
-                }
-            });
-        }
-
-        function goBack() {
-            location.href = '/student';
-        }
-    </script>
-</body>
-</html>
